Adjust default setup to something slightly more tractable
This by default involves just trunk and a single stable version, which I
have updated to be 0.24. (For the sake of example it might be better to
have two stable versions, but I won't do that right now.)

I've tweaked a range of other defaults to more useful testing values,
and added some comments.

This commit also updates the webtiles starting scripts to use python3 by
default.