Add a docker build process for a testing container
This commit introduces a system for automating the construction of a
docker container that corresponds to:

    https://crawl.develz.org/wiki/doku.php?id=dcss:server:setting_up_dgamelaunch_and_webtiles

This is aimed at providing more robust tools for testing and maintaining
existing dgamelaunch-config setups, rather than any sort of
containerization for production purposes.

There are also a few improvements to the core scripts in this commit;
more directories are automatically created, and there is a script in
utils to build a fresh chroot environment. (Be aware that this script is
not minimalist in terms of what gets installed.)

(cherry picked from commit cca4647e65ff4f45070729257be2522484e99fd0)