Update configuration for 0.25
We don't necessarily have to update the dgamelaunch config every
release, but I was testing this repo and thought I would.

The supplied RC file was quite out of date, referring to options removed
a long time ago. It might be good to update the scripts to not have a
copy of the RC and get a copy from the appropriate repo at some point
(publishing?) if a local file hasn't been made by an admin. In practice,
I think all servers use the default RC from the repo, but we want to
continue to allow having a version with server-specific modifications in
the chroot.

Additionally tweak the menus to give the latest released version a
dedicated key on 'R' so we don't have to update the key by number with
every release.

I've removed some stale references to old versions in various places.
The rebuild script is awkward, since it hard-codes viable versions and a
pattern to match them, something we should try to refactor.