Allow installation of branches via stable scripts
The stable install/update scripts support installing some alternate
branch by specifying an argument, which would allow easy installation
and updating of experimental branches on the crawl side of things.
However there's currently a regex test against the version argument to
make sure it's a version number that prevents use of this feature. This
commit disables this regex test, allowing the scripts to work when given
a branch as an argument. Doing so doesn't introduce any problems I'm
aware of, since version number arguments are still handled properly, and
if any specified branch doesn't exist, the script will still error when
it tries to do the git checkout.

Ideally installation of special branches would be handled by a seperate
script, at least in terms of evocation; these branches are generally not
anything close to "stable".