fix: Don't try to publish RC or macro files
The dgl scripts try to publish RC and macro files stored in this repo to
allow servers to have custom defaults. I don't believe any DGL servers
use this functionality, and it results in publishing potentially stale
files. Simply removing the in-repo files will cause errors with the
publishing script, so this commit also removes support for publishing
custom RC and macro files in the script.

The existing dgamealaunch config already uses the RC files installed by
the installer rather than those published by the dgl scripts. For
macros, we tweak the commands for resetting options to simply unlink the
user's macro file instead of copying a default file.