feat: Scripts to support experimental branches
Previously we updated the stable branch scripts to allow them to work
with experimentals, but this had unwanted side effects. Directories were
created that experimentals don't use, some of which mangled the branch
name into a numeric version, and there was no support for using a game
name different from the git branch name. The latter is important since
developers like to use funny names that might confuse users in e.g.
WebTiles lobby listings.

This commit adds a set of scripts specifically for experimentals: a
build updater, an installer, and a launcher. The updater require two
arguments, a git branch name used for git purposes, and a game name used
in naming the crawl binary and any relevant directories.  The installer
and launcher only expect the game name, since that's the only value they
need. The installer doesn't create sprint or tutorial directories, since
we don't use those for experimentals. The launcher uses the user's git
RC file for its RC.

We also revert changes to the stable updater that relaxed conditions on
the branch name. This script goes back to being specialized for the
stable DCSS release branches.