feat: A bot-only crawl install
For running bots like qw, its helpful to have installs of stable and
trunk that are accessible only to bot accounts. This way qw can run on
an untracked server like CDI, but the bot-only logfiles and milestone
files can still be added to Sequell, Tournament, and CAO scoring.  In
the future, bots might additionally need slightly different game build
options.

This commit adds a bot user flag (value 64) and updaters, installers,
and launchers for bot-only installs of stable and trunk. The launcher
refuses to execute crawl if the user does not have the bot flag,
providing an appropriate message in both dgamelaunch and webtiles when
this happens. When running the bot stable/git updaters along with the
regular updaters, a proper ccache setup will ensure that the bot builds
are very fast. Both installers do the same type of install as the
regular stable installer, namely games are always forced transfered.
Bots don't need to decide on a per-commit basis whether to transfer a
trunk game, and not maintaining an additional set of per-commit builds
of trunk simplifies management.

Since this functionality is likely only useful to myself and CDI, I'm
not adding any configuration examples to dgamelaunch.conf or configy.py.