feat: support for new webtiles config files
This adds explicit templating and file management support for
config.yml, banned names files, and games.d files, as well as an
override for the default games.d/base.yaml file (though this override
isn't needed for the default configuration).  No doubt the perl here
isn't very fluent, but I've tested the basics out on cao and it seems to
work.