feat: Support force pushes to git branches
Currently we do a `git pull` to update the repository, which of course
doesn't work in the case where the branch has been force pushed. Force
pushes shouldn't happen (hopefully) for stable or trunk, but they're
common for experimental branches. This commit has the repo update do a
`git fetch` followed by a hard reset to the relevant branch. It
additionally updates the stable installer to remove the data directory
with each installation. This way we'll not have stale data files lying
around and potentially loaded by crawl if they get removed from the
repository by a previous commit.